added option for account deletion requests

+++ webroot/panel/account.php
@@ -70,6 +70,16 @@ if ($_SERVER['REQUEST_METHOD'] == "POST") {
                 }
             }
             break;
+        case "account_deletion_request":
+            $hasGroups = count($USER->getGroups()) > 0;
+            if ($hasGroups) {
+                die();
+                break;
+            }
+            if (!$SQL->accDeletionRequestExists($USER->getUID())) {
+                $USER->requestAccountDeletion();
+            }
+            break;
     }
 }
 ?>
@@ -107,17 +117,30 @@ if ($isPI) {
 }
 
 if (!$isPI) {
-    echo
-    "<form action='' method='POST' id='piReq' 
-    onsubmit='return confirm(\"Are you sure you want to request a PI account?\")'>
-    <input type='hidden' name='form_type' value='pi_request'>";
-    if ($SQL->requestExists($USER->getUID())) {
+    if ($SQL->accDeletionRequestExists($USER->getUID())) {
+        echo
+        "<form action='' method='POST' id='piReq' 
+        onsubmit='return confirm(\"Are you sure you want to request a PI account?\")'>
+        <input type='hidden' name='form_type' value='pi_request'>";
         echo "<input type='submit' value='Request PI Account' disabled>";
-        echo "<label style='margin-left: 10px'>Your request has been submitted and is currently pending</label>";
+        echo
+        "<label style='margin-left: 10px'>
+            You cannot request PI Account while you have requested account deletion.
+        </label>";
+        echo "</form>";
     } else {
-        echo "<input type='submit' value='Request PI Account'>";
+        echo
+        "<form action='' method='POST' id='piReq' 
+        onsubmit='return confirm(\"Are you sure you want to request a PI account?\")'>
+        <input type='hidden' name='form_type' value='pi_request'>";
+        if ($SQL->requestExists($USER->getUID())) {
+            echo "<input type='submit' value='Request PI Account' disabled>";
+            echo "<label style='margin-left: 10px'>Your request has been submitted and is currently pending</label>";
+        } else {
+            echo "<input type='submit' value='Request PI Account'>";
+        }
+        echo "</form>";
     }
-    echo "</form>";
 }
 ?>
 
@@ -193,6 +216,33 @@ for ($i = 0; $sshPubKeys != null && $i < count($sshPubKeys); $i++) {  // loop th
     <input type='submit' value='Set Login Shell'>
 
 </form>
+
+<hr>
+
+<h5>Account Deletion</h5>
+<?php
+$hasGroups = count($USER->getGroups()) > 0;
+
+if ($hasGroups) {
+    echo "<p>You cannot request to delete your account while you are in a PI group.</p>";
+} else {
+    echo
+    "<form action='' method='POST' id='accDel' 
+    onsubmit='return confirm(\"Are you sure you want to request an account deletion?\")'>
+    <input type='hidden' name='form_type' value='account_deletion_request'>";
+    if ($SQL->accDeletionRequestExists($USER->getUID())) {
+        echo "<input type='submit' value='Request Account Deletion' disabled>";
+        echo "<label style='margin-left: 10px'>Your request has been submitted and is currently pending</label>";
+    } else {
+        echo "<input type='submit' value='Request Account Deletion'>";
+    }
+    echo "</form>";
+}
+
+?>
+
+<hr>
+
 
 <script>
     $("button.btnAddKey").click(function() {
